+++ main/modes/ray/ray_object.h
@@ -5,7 +5,7 @@
 
 void initEnemyTemplates(ray_t* ray);
 void fireShot(ray_t* ray);
-void moveRayObjects(ray_t* ray, int64_t elapsedUs);
+void moveRayObjects(ray_t* ray, int32_t elapsedUs);
 void checkRayCollisions(ray_t* ray);
 
 #endif
